Implemented sort fields for GetAllAsync() and Test campaign with email in a test request (#38)
* Get all campaigns with sort order "send_time" or "create_time"

* MailChimpException suppress more exception sources and Campaign test with email as RequestContent

* Campaign ScheduleAsync and request object content

* Type for _rate proprties changed to double and added missing enums for ConditionTypes and Operators

+++ MailChimp.Net/Models/Clicks.cs
@@ -17,7 +17,7 @@ namespace MailChimp.Net.Models
         /// Gets or sets the click rate.
         /// </summary>
         [JsonProperty("click_rate")]
-        public int ClickRate { get; set; }
+        public double ClickRate { get; set; }
 
         /// <summary>
         /// Gets or sets the clicks total.
